feat!: remove dependency on pythonocc
Until now, klingo was using pythonocc-core v7.5.1 to export surfaces to
STL files. This was only a temporary solution as pythonocc requires
many dependencies and, thus, does not fit a 'minimal' framework.

Now, the Blade class implements the `export_stl` method, which uses
numpy-stl behind the scenes to export the blade surface to STL.